Create openfile.go
Move preprocessor of file to new file.
